spike: npm support with modernization
The purpose of this spike is to allow
projects to import quietjs through npm.
In doing so the code has been split
apart into modules, although all logic
related to transmitting and receiving
should be the same.

This effort was undertaken with
coordination with the maintainer of
quietijs, but this branch has been
abandoned and the effort given to
Gabriel to complete.